null array

+++ BizServer/service/model.go
@@ -8,13 +8,14 @@ import (
 type User struct {
 	Name      string `json:"name"`
 	Family    string `json:"family"`
-	Id        int64  `json:"id"`
-	Age       int64  `json:"age"`
+	Id        int32  `json:"id"`
+	Age       int32  `json:"age"`
 	Sex       string `json:"sex"`
 	CreatedAt int64  `json:"createdAt"`
 }
 
 func (p *User) getUser(db *sql.DB) error {
+
 	return errors.New("Not implemented")
 }
 func (p *User) updateUser(db *sql.DB) error {
